feat(core): split commands.ts into domain modules
Fixes #305
