+++ .squad/agents/fenster/history.md
@@ -55,6 +55,15 @@ DevDocket is a VS Code extension monorepo for managing work items from multiple 
 
 ## Learnings
 
+### 2026-04-23 — Issue #305 (Split commands.ts into domain modules)
+
+**Refactoring:** Split the 1118-line monolith `commands.ts` into 8 domain-specific modules plus a shared utilities file.
+- **Modules created:** `commandUtils.ts` (shared helpers), `inboxCommands.ts`, `queueCommands.ts`, `focusCommands.ts`, `historyCommands.ts`, `layoutCommands.ts`, `generalCommands.ts`, `sourcesCommands.ts`, `watchCommands.ts`.
+- **Pattern:** Each module exports a single `register*Commands()` function that receives only the dependencies it needs. The original `commands.ts` becomes a thin orchestrator calling each domain registrar.
+- **Shared utilities in `commandUtils.ts`:** `wrapCommand`, `handleCommandError`, `resolveItemIds`, `formatItemTitle`, `batchTransition`, `batchAcceptItems` + `AcceptableItem` interface — used across multiple domain modules.
+- **Key lesson:** When splitting a monolith, identify cross-cutting helpers first and extract them into a shared utils module. Domain-specific type guards (e.g., `isInboxItem`, `isSourceItem`) stay in their respective domain modules since they're only used there.
+- **Files changed:** 9 new files in `packages/core/src/commands/`, `commands.ts` reduced to ~40 lines.
+
 ### 2026-04-22 — Issue #300 (CancellationToken → AbortSignal wiring)
 
 **Bug fix:** Providers accepted `CancellationToken` in `refresh()` but only checked `isCancellationRequested` at discrete points. In-flight `fetch()` calls ran to completion even after cancellation.
